#283 move: 임시 메뉴 삭제하고 세부 통계 페이지로 코드 이동
