+++ src/pages/DashboardPage/CompletionChart.jsx
