+++ src/pages/DashboardPage/MajorChart.jsx
