--- src/pages/DetailStatisticsPage/index.js
@@ -1,1 +1,0 @@
-export { default as DetailStatisticsPage } from './DetailStatisticsPage';
